Aprimoraa funcao de actualizacao dos dados do usuario

+++ app/Repositories/Customer/CustomerRepository.php
@@ -25,7 +25,9 @@ class CustomerRepository
         $user = $this->getAuthUser();
 
         if (isset($data['photo'])) {
-            Storage::disk('public')->delete($user->photo_path);
+            if($user->photo_path){
+                Storage::disk('public')->delete($user->photo_path);
+            }
             $image = $data['photo'];
             $path =  $image->store('avatar', 'public');
             $data['photo_path'] = $path;
